refactor(fleet-core): extract the provider quota stack into core-ai-gateway
