+++ packages/core-ai-gateway/tests/quota/service.test.ts
@@ -1,7 +1,9 @@
 import { describe, expect, it, vi } from "vitest";
 
-import { createQuotaService } from "../server/service.js";
-import type { ProviderSuccess } from "../server/types.js";
+import { createAiGatewayQuotaCollectors, createQuotaService } from "../../src/quota/service.js";
+import { KIMI_AUTH_PROVIDER_ID } from "../../src/models.js";
+import { OPENCODE_AUTH_PROVIDER_ID } from "../../src/opencode-go/index.js";
+import type { ProviderSuccess } from "../../src/quota/types.js";
 
 function ok(fetchedAt: number, usedPercent = 10): ProviderSuccess {
   return { status: "ok", fetchedAt, windows: [{ id: "session", usedPercent }] };
@@ -18,6 +20,7 @@ describe("quota service", () => {
       fetchClaude,
       fetchCodex: async () => ({ status: "signed_out" }),
       fetchCursor: async () => ({ status: "signed_out" }),
+      fetchOpencode: async () => ({ status: "signed_out" }),
     });
     expect((await service.getSummary()).providers.claude).toEqual({ status: "not_connected", method: "keychain" });
     expect(fetchClaude).not.toHaveBeenCalled();
@@ -33,6 +36,7 @@ describe("quota service", () => {
       fetchClaude: async () => ({ status: "signed_out" }),
       fetchCodex: async () => ({ status: "signed_out" }),
       fetchCursor,
+      fetchOpencode: async () => ({ status: "signed_out" }),
     });
     expect((await service.getSummary()).providers.cursor)
       .toEqual({ status: "not_connected", method: "keychain" });
@@ -51,6 +55,7 @@ describe("quota service", () => {
       fetchClaude,
       fetchCodex: async () => ({ status: "signed_out" }),
       fetchCursor: async () => ({ status: "signed_out" }),
+      fetchOpencode: async () => ({ status: "signed_out" }),
     });
     const first = service.getSummary();
     const second = service.getSummary();
@@ -83,6 +88,7 @@ describe("quota service", () => {
       fetchClaude,
       fetchCodex,
       fetchCursor,
+      fetchOpencode: async () => ({ status: "signed_out" }),
     });
     const cached = await service.getSummary();
     const refreshed = await service.getSummary({ forceProvider: "cursor" });
@@ -92,6 +98,29 @@ describe("quota service", () => {
     expect(refreshed.providers.claude).toEqual(cached.providers.claude);
     expect(refreshed.providers.codex).toEqual(cached.providers.codex);
     expect(refreshed.providers.cursor.windows?.[0]?.usedPercent).toBe(32);
+  });
+
+  it("collectors read Kimi and OpenCode keys only through the injected auth service", async () => {
+    const authService = {
+      getApiKey: vi.fn(async (providerId: string) => providerId === KIMI_AUTH_PROVIDER_ID ? "kimi-key" : "opencode-key"),
+      setApiKey: async () => undefined,
+      deleteApiKey: async () => false,
+      listProviderIds: async () => [],
+    };
+    const fetchImpl = vi.fn(async (url: string | URL | Request) => new Response(
+      String(url).includes("kimi") ? JSON.stringify({ usage: { limit: "100", used: "4" } }) : "{}",
+      { status: 200, headers: { "Content-Type": "application/json" } },
+    ));
+    const collectors = createAiGatewayQuotaCollectors({
+      authService,
+      fetch: fetchImpl as typeof fetch,
+      scanOpencodeWindows: async () => null,
+    });
+
+    await expect(collectors.fetchKimi()).resolves.toMatchObject({ status: "ok" });
+    await expect(collectors.fetchOpencode()).resolves.toMatchObject({ status: "ok" });
+    expect(authService.getApiKey).toHaveBeenNthCalledWith(1, KIMI_AUTH_PROVIDER_ID);
+    expect(authService.getApiKey).toHaveBeenNthCalledWith(2, OPENCODE_AUTH_PROVIDER_ID);
   });
 
   it("serves last-good data as stale for 30 minutes, then returns sanitized error", async () => {
@@ -107,6 +136,7 @@ describe("quota service", () => {
       fetchClaude,
       fetchCodex: async () => ({ status: "signed_out" }),
       fetchCursor: async () => ({ status: "signed_out" }),
+      fetchOpencode: async () => ({ status: "signed_out" }),
     });
     await service.getSummary();
     now += 1_799_999;
